Solved problems of unable to add/remove/update computer in the inventory
I set inventory attribute as a dictionary in the first time so I wrote the methods accordingly. Last time I got from the feedback it should be a list so I changed that but forgot to update the methods. This time I fixed the problems and added a few more test cases. Hopefully it will work!